Included single module and all module support

+++ src/records/RecordsDelete.java
@@ -10,16 +10,21 @@ public class RecordsDelete {
 	
 	private boolean allRecords;
 	private boolean allModules;
-	public RecordsDelete(boolean allModules, boolean allRecords)
+	private String[] modules;
+	public RecordsDelete(boolean allModules, boolean allRecords, String[] modules)
 	{
 		this.allRecords = allRecords;
 		this.allModules = allModules;
+		if(!allModules)
+		{
+			this.modules = modules;
+		}
 	}
 	public void deleteAllRecords(String module) throws Exception
 	{
 		ArrayList<String> idParamList = new ArrayList<String>();
 		ArrayList<String> ids = new ArrayList<String>();
-		int maxCount = 100;
+		int maxCount = 700;
 		Integer page = 1;
 		while(maxCount>0)
 		{
@@ -35,13 +40,22 @@ public class RecordsDelete {
 			}
 			else
 			{
+				int sc = response.getStatusCode();
+				System.out.println("\n\nResponse code: "+sc);
+				if(sc!=204)
+				{
+					System.out.println("\n\n"+response.getResponse().toString()+"\n\n\n");
+				}
 				break;
 			}
 			maxCount--;
 		}
+//		System.out.println("\n\nSystem going to sleep for sometime\n\n");
+//		Thread.sleep(5000);
 		int totalLen = ids.size();
 		int iter = totalLen/99;
 		int mod = totalLen%99;
+		ArrayList<Integer> lens = new ArrayList<Integer>();
 		for(int i=1; i<=iter; i++)
 		{
 			int maxIndex = i*99;
@@ -49,6 +63,7 @@ public class RecordsDelete {
 			List<String> idList = ids.subList(minIndex, maxIndex);
 			String idString = String.join(",", idList);
 			idParamList.add(idString);
+			lens.add(idList.size());
 		}
 		if(mod>0)
 		{
@@ -57,12 +72,25 @@ public class RecordsDelete {
 			List<String> idList = ids.subList(minIndex, maxIndex);
 			String idString = String.join(",", idList);
 			idParamList.add(idString);
+			lens.add(idList.size());
 		}
+		int deleted = 0;
+		int delIndex = 0;
 		for(String idParam: idParamList)
 		{
 			RestApiRequest delRequest = new RestApiRequest(RestApiRequest.DELETE,module);
 			delRequest.addParam("ids", idParam);
-			delRequest.executeRequest();
+			RestApiResponse resp = delRequest.executeRequest();
+			if(resp.getStatusCode()==200 && delIndex<lens.size())
+			{
+				deleted+=lens.get(delIndex);
+				System.out.println("\n\n"+deleted+" out of "+totalLen+" deleted");
+				delIndex++;
+			}
+			else
+			{
+				System.out.println("\n\nThere was a problem while trying to delete the records!");
+			}
 		}
 	}
 	public void deleteRecords() throws Exception
@@ -78,14 +106,41 @@ public class RecordsDelete {
 			}
 			if(moduleList!=null)
 			{
+				int inventoryCount = 0;
+				boolean productsDone = false;
 				for(String module: moduleList)
 				{
+					if(module.equals("Products"))
+					{
+						continue;
+					}
 					if(this.allRecords)
 					{
 						this.deleteAllRecords(module);
 					}
+					if(module.equals("Quotes") || module.equals("Sales_Orders") || module.equals("Purchase_Orders") || module.equals("Invoices"))
+					{
+						inventoryCount++;
+					}
+					if(inventoryCount==4 && !productsDone)
+					{
+						this.deleteAllRecords("Products");
+						productsDone=true;
+					}
 				}
 			}
+		}
+		else if(this.modules!=null)
+		{
+			for(int i=0; i<this.modules.length; i++)
+			{
+				String module = modules[i];
+				if(this.allRecords)
+				{
+					this.deleteAllRecords(module);
+				}
+			}
+			
 		}
 	}
